Fix an issue in a special case when replacing points.
Using remainder (%) on a value is, admittedly, defensive programming. But it's very useful when using arrays and ranges that can wrap beyond the ends.

The consequences are that, when the location=_count and length=0, we don't use remainder, so that new points are appended to the list. Otherwise, if length > 0, new remove/replace points at the beginning of the list.