Use balanced bias when update border nets.
This fixes a particular problem with updating the border nets.

If we have points [a, b, c, ... y, z], and yza makes a triangle that has two segments, and we then add d such that we get [d, c, .., y, z], .
Then the hull segments become [Sab, Sbc, .. Syz, Sza] -> [Sdc, .. Syz, Szd]. I.E. we remove two segments from the front, and one from and end of list, and add a new one at the front and end.

d is added at the front by the range calculator, and the segments are derived. But the triangle is in the border net will fail with the original range calculation behaviour. It will remove one triangle and the end and replace two at the front, whereas we want it to replace one at the end and substitute 1 new for 2 old at the front.

By changing the bias in the case, we get the correct behaviour. This may not solve all such array modification problems. We may have to use a more explicit and introspective approach, matching up boundary triangles with points/segments in hull.